requirements.txt
fastapi[all]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
ldap3
uvicorn
sqlmodel
alembic
jinja2
python-multipart
aiofiles
pytest
pytest-env
black
invoke
ipython
requests
chardet
